Fix path truncation in drag&drop on WindowsDX
This fixes an off-by-one error in WindowsDX drag&drop handling that caused
paths of dropped files to be truncated by one character.

This happens due do a mismatch between DragQueryFile returning the number of
characters in a path that it will write but expecting to be passed the buffer
size including null terminator.